# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: RankMap API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: clusterId
          in: query
          required: false
          schema:
            type: integer
        - name: intent
          in: query
          required: false
          schema:
            type: string
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordBody"
      responses:
        "201":
          description: Created keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /projects/{projectId}/keywords/import:
    post:
      operationId: importKeywords
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportKeywordsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportKeywordsResponse"
  /projects/{projectId}/keywords/{id}:
    get:
      operationId: getKeyword
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateKeyword
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKeywordBody"
      responses:
        "200":
          description: Updated keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/score-settings:
    get:
      operationId: getScoreSettings
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreSettings"
    patch:
      operationId: updateScoreSettings
      tags: [keywords]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScoreSettingsBody"
      responses:
        "200":
          description: Updated score settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Keyword:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        tenantId:
          type: integer
        phrase:
          type: string
        searchVolume:
          type: ["integer", "null"]
        cpc:
          type: ["number", "null"]
        kd:
          type: ["integer", "null"]
        intent:
          type: ["string", "null"]
        clusterId:
          type: ["integer", "null"]
        source:
          type: string
          enum: [manual, csv, ahrefs, semrush]
        isActive:
          type: boolean
        rawScore:
          type: ["number", "null"]
        finalScore:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - tenantId
        - phrase
        - source
        - isActive
        - createdAt
        - updatedAt
    CreateKeywordBody:
      type: object
      properties:
        phrase:
          type: string
        searchVolume:
          type: integer
        cpc:
          type: number
        kd:
          type: integer
        intent:
          type: string
          enum: [informational, navigational, commercial, transactional]
      required:
        - phrase
    UpdateKeywordBody:
      type: object
      properties:
        phrase:
          type: string
        searchVolume:
          type: integer
        cpc:
          type: number
        kd:
          type: integer
        intent:
          type: string
          enum: [informational, navigational, commercial, transactional]
        isActive:
          type: boolean
        clusterId:
          type: integer
    ImportKeywordsBody:
      type: object
      properties:
        source:
          type: string
          enum: [manual, csv, ahrefs, semrush]
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/RawKeywordInput"
      required:
        - source
        - keywords
    RawKeywordInput:
      type: object
      properties:
        phrase:
          type: string
        searchVolume:
          type: integer
        cpc:
          type: number
        kd:
          type: integer
        intent:
          type: string
      required:
        - phrase
    ImportKeywordsResponse:
      type: object
      properties:
        imported:
          type: integer
        duplicates:
          type: integer
        total:
          type: integer
      required:
        - imported
        - duplicates
        - total
    ScoreSettings:
      type: object
      properties:
        projectId:
          type: integer
        volumeWeight:
          type: number
        kdWeight:
          type: number
        intentWeight:
          type: number
        cpcWeight:
          type: number
        freshnessWeight:
          type: number
      required:
        - projectId
        - volumeWeight
        - kdWeight
        - intentWeight
        - cpcWeight
        - freshnessWeight
    UpdateScoreSettingsBody:
      type: object
      properties:
        volumeWeight:
          type: number
        kdWeight:
          type: number
        intentWeight:
          type: number
        cpcWeight:
          type: number
        freshnessWeight:
          type: number
...